+++ src/confluent_kafka/src/Consumer.c
@@ -1142,32 +1142,41 @@ Consumer_consumer_group_metadata (Handle *self, PyObject *ignore) {
 static PyObject *
 Consumer_io_event_enable(Handle *self, PyObject *args, PyObject *kwargs)
 {
-    int fd = 0;
-    const char* data = NULL;
-    Py_ssize_t data_size = 0;
-    static char *kws[] = {"fd", "payload", NULL};
+        int fd = 0;
+        const char* data = NULL;
+        Py_ssize_t data_size = 0;
+        static char *kws[] = {"fd", "payload", NULL};
 
-	if (!PyArg_ParseTupleAndKeywords(args, kwargs, "iy#", kws,
-                &fd, &data, &data_size))
-    {
-		return NULL;
-    }
+        if (!self->rk) {
+                PyErr_SetString(PyExc_RuntimeError, "Consumer closed");
+                return NULL;
+        }
 
-    if (fd <= 0) {
-        return PyErr_Format(PyExc_ValueError, "fd outside range: %i", fd);
-    }
+        if (!PyArg_ParseTupleAndKeywords(args, kwargs, "iy#", kws, &fd,
+                                         &data, &data_size)) {
+                return NULL;
+        }
 
-    if (data_size <= 0) {
-        return PyErr_Format(PyExc_ValueError, "payload cannot be empty");
-    }
+        if (fd < 0) {
+                return PyErr_Format(PyExc_ValueError,
+                                    "fd outside range: %i", fd);
+        }
 
-    if (self->u.Consumer.rkqu == NULL) {
-        PyErr_SetString(PyExc_RuntimeError, "Consumer Queue not available");
-        return NULL;
-    }
+        if (data_size <= 0) {
+                return PyErr_Format(PyExc_ValueError,
+                                    "payload cannot be empty");
+        }
 
-    rd_kafka_queue_io_event_enable(self->u.Consumer.rkqu, fd, data, data_size);
-    Py_RETURN_NONE;
+        if (self->u.Consumer.rkqu == NULL) {
+                PyErr_SetString(PyExc_RuntimeError,
+                                "Consumer Queue not available");
+                return NULL;
+        }
+
+        rd_kafka_queue_io_event_enable(self->u.Consumer.rkqu,
+                                       fd, data, data_size);
+
+        Py_RETURN_NONE;
 }
 
 static PyMethodDef Consumer_methods[] = {
@@ -1525,22 +1534,19 @@ static PyMethodDef Consumer_methods[] = {
         { "set_sasl_credentials", (PyCFunction)set_sasl_credentials, METH_VARARGS|METH_KEYWORDS,
            set_sasl_credentials_doc
         },
-        {
-            "io_event_enable",
-            (PyCFunction) Consumer_io_event_enable,
-            METH_VARARGS | METH_KEYWORDS,
-            ".. py:function:: io_event_enable(fd, payload)\n"
-            "\n"
-            "This method basically calls ``rd_kafka_queue_io_event_enable``\n"
-            "\n"
-            "Enable IO event triggering.\n"
-            "To ease integration with IO based polling loops this API\n"
-            "allows an application to create a separate file-descriptor\n"
-            "that librdkafka will write ``payload`` to whenever a new\n"
-            "element is enqueued on a previously empty queue.\n"
-            " :param int fd: The filedescrptor librdkafka writes to.\n"
-            " :param bytes payload: The payload librdkakfa writes to fd.\n"
-            " :returns: ``None``\n"
+        { "io_event_enable",
+          (PyCFunction) Consumer_io_event_enable,
+          METH_VARARGS | METH_KEYWORDS,
+          ".. py:function:: io_event_enable(fd, payload)\n"
+          "\n"
+          "Enable IO event triggering.\n"
+          "To ease integration with IO based polling loops this API\n"
+          "allows an application to create a separate file-descriptor\n"
+          "that the consumer will write ``payload`` to whenever a new\n"
+          "element is enqueued on a previously empty queue.\n"
+          " :param int fd: The filedescriptor the consumer writes to.\n"
+          " :param bytes payload: The payload the consumer writes to fd.\n"
+          " :returns: ``None``\n"
         },
 
 	{ NULL }